Added movement controller that will slow down as it nears the target. This will help reduce the overshoot due to robot momentum. The parameters of when and how much to slow down might need altering somewhat -  i.e. maybe start slowing down earlier, or slowing down even more than 0.5x speed.
Also, when it reaches the target it will wait for some time (100ms
currently) for the robot to settle, and then re-enable the controller
for a short time (150ms currently). This will wait for any overshoot to
happen (which should be less now anyway) and then start moving again at
the slow speed to better align. The time to settle may need altering. If
the slow down code works then the additional re-enable of the controller
might not do much, but it's only 0.25 of a second. Depending on
performance the re-enable may need even more slow down, like 0.1 during
re-enable period.